+++ src/components/atoms/FriendsList/FriendsListB/index.tsx
@@ -1,12 +1,14 @@
-import { FlexBox } from '@components/layouts/FlexBox'
-import { Text } from '@components/atoms/Text'
 import styled from '@emotion/styled'
-import { theme } from '@styles/theme'
-import { Avatar } from '@components/atoms/Avatar'
-import Svg from '@components/atoms/Svg'
-import CircleIcon from '@assets/icons/CircleIcon'
-import { sliceString } from '@utils/sliceString'
+
 import { palette } from '@styles/theme/palette'
+import { FlexBox } from '@components/layouts/FlexBox'
+import { Avatar } from '@components/atoms/Avatar'
+import {
+  FriendsListBPropsType,
+  DescriptionType,
+} from '@models/components/atoms/FriendsList'
+import FriendsDescription from './FriendsDescription'
+import FriendsCurrentState from './FriendCurrentState'
 
 /**
  * @param userName 친구 이름을 나타냄
@@ -18,18 +20,6 @@ import { palette } from '@styles/theme/palette'
  * @param birthday 생일 일자를 나타냄
  */
 
-export type DescriptionType = 'birthday' | 'none' | 'newUpdate'
-
-export type FriendsListBProps<T extends DescriptionType> = {
-  userName: string
-  currentState: string
-  onClick?: () => void
-  imageUrl: string
-  description: T
-  birthdayDescription?: T extends 'birthday' ? string : undefined
-  birthday?: T extends 'birthday' ? string : undefined
-}
-
 const FriendsListB = ({
   userName,
   currentState,
@@ -38,46 +28,19 @@ const FriendsListB = ({
   birthday,
   birthdayDescription,
   ...props
-}: FriendsListBProps<DescriptionType>) => {
+}: FriendsListBPropsType<DescriptionType>) => {
   return (
     <Wrapper {...props}>
       <FriendsProfileWrapper>
         <Avatar variant="small" imageUrl={imageUrl} />
-        <FriendsInfoWrapper>
-          <Text typo="Subhead_14" color="white">
-            {userName}
-          </Text>
-          {description === 'birthday' && (
-            <FlexBox justify="flex-start" gap={2}>
-              <Text color="gray_300" typo="Caption_10">
-                {birthdayDescription && birthdayDescription + ' · '}
-              </Text>
-              <Text color="gray_300" typo="Caption_10">
-                {birthday}
-              </Text>
-            </FlexBox>
-          )}
-          {description === 'newUpdate' && (
-            <FlexBox justify="flex-start" gap={2}>
-              <Svg children={<CircleIcon />} style={{ display: 'flex' }} />
-              <Text color="purple_400" typo="Caption_10">
-                새로운 업데이트
-              </Text>
-            </FlexBox>
-          )}
-        </FriendsInfoWrapper>
+        <FriendsDescription
+          description={description}
+          userName={userName}
+          birthdayDescription={birthdayDescription}
+          birthday={birthday}
+        />
       </FriendsProfileWrapper>
-      <FriendsCurrentStateWrapper>
-        <StyledTextWrapper length={currentState.length > 15 ? 1 : 0}>
-          <Text
-            typo="Caption_10"
-            color="gray_200"
-            as="div"
-            style={{ width: currentState.length > 15 ? '100%' : 'fit-content' }}
-            children={sliceString(currentState, 15)}
-          />
-        </StyledTextWrapper>
-      </FriendsCurrentStateWrapper>
+      <FriendsCurrentState currentState={currentState} />
     </Wrapper>
   )
 }
@@ -97,26 +60,4 @@ const FriendsProfileWrapper = styled(FlexBox)`
   gap: 12px;
   width: 180px;
   height: 100%;
-`
-
-const FriendsInfoWrapper = styled(FlexBox)`
-  flex-direction: column;
-  gap: 2px;
-  align-items: flex-start;
-  width: 100%;
-  height: 100%;
-`
-
-const FriendsCurrentStateWrapper = styled(FlexBox)`
-  width: 144px;
-  height: 100%;
-  justify-content: flex-end;
-`
-
-const StyledTextWrapper = styled(FlexBox)<{ length: 0 | 1 }>`
-  width: ${({ length }) => (length ? `100%` : `fit-content`)};
-  height: 26px;
-  padding: 6px 8px;
-  border-radius: 13px;
-  background-color: ${theme.palette.gray_800};
 `
